move ResourceHelper into a test/ directory
